src/main/java/com/capgemini/adm/tdd/shapes/Circle.java
package com.capgemini.adm.tdd.shapes;

public final class Circle {
    private final double radius;

    public Circle(double radius) {
        this.radius = radius;
    }

    public double getRadius() {
        return radius;
    }

    public double getDiameter() {
        return 2 * radius;
    }

    public double getArea() {
        return Math.PI * radius * radius;
    }

    public double getCircumference() {
        return Math.PI * 2.0 * radius;
    }

    public int getSides() {
        return 0;
    }
}